requirements.txt
Django==4.1
djangorestframework==3.13.1
psycopg2>=2.8.6,<2.9
django-phonenumber-field==7.1
phonenumbers==8.13
drf-spectacular==0.24.1